Separate module and type
